Changes to reducer structure
changed list reducer to now hold an array of item ids, and the item reducer no longer tracks index. This is in preparation
for drag-to-reorder, as re-ordering/inserting into arrays makes more sense than changing an index value on the item objects.

Also is more reasonable in the sense that an item's index in a list is something that the list needs to know about, but not the
item, so it doesn't make sense to keep that information on the item object.